+++ packages/evokit-box/CHANGELOG.md
@@ -9,3 +9,5 @@
 #### 3.0.0
 
 - Initial version independent package. The [previous changelog](packages/evokit/CHANGELOG.md) in another package.
+- **Added** new prop `box-height`
+- **Added** new prop `box-overflow`